playing with enums and impl of organism trait, added some tests

+++ src/utils/domain.rs
@@ -1,7 +1,8 @@
 use rand::Rng;
 
-pub fn random_float(low: f32, high: f32) -> f32 {
+pub fn random_float(low: f64, high: f64) -> f64 {
         let mut rng = rand::thread_rng();
-        let n: f32 = rng.gen_range(low..=high);
+        let n: f64 = rng.gen_range(low..=high);
         n
 }
+
